requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.13.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.1
urllib3==1.26.6
whitenoise==5.2.0
